+++ src/fiber_pool.hpp
@@ -14,15 +14,15 @@ namespace big
 		void fiber_tick();
 		static void fiber_func();
 
-		int get_total_fibers();
-		int get_used_fibers();
+		size_t get_total_fibers();
+		size_t get_used_fibers();
 
 		void reset();
 
 	private:
 		std::recursive_mutex m_mutex;
 		std::stack<std::function<void()>> m_jobs;
-		int m_num_fibers;
+		size_t m_num_fibers;
 	};
 
 	inline fiber_pool* g_fiber_pool{};
